pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bip-bolt-report"
version = "0.1.0"
description = "Analyses token counts and LLM cost-to-explain across all Bitcoin and Lightning standards."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Mohit-Davar <mohitdavar2004@gmail.com>" }]

dependencies = [
    "gitpython>=3.1",
    "tiktoken>=0.7",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "click>=8.1",
    "pandas>=2.0",
    "rich>=13.0",
    "mwparserfromhell>=0.6",
]

[project.scripts]
generate = "report.__main__:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
